Add examples
These examples work not only as examples for new users, but also
as a check that any change does not break the existing workflow.
They are to be considered as tests, in that a change to them indicates
a breaking change in the project as a whole.

Writing proper unit tests is a goal. Due to the nature of proc-macros
and in particular the hard to work with nature of the syn crate, unit
tests will have to be developed sometime in the future.