Prepare toml files for crates.io publish
The rust version was determined by the dependency to regex, the main
deadlocker crate could probably work with a lower rust version. However,
since the main point of this project is the derive macro I have set the
rust version in the main crate to match that of the derive crate.